Make better use of context-cached "immediates".
This commit is a global search and replace of all code that
traverses through context.runtime to get at nil, true, and false
objects rather than using the cached versions on context itself.
The context is thread-local and probably very likely to be in the
CPU cache, which means these references should be in cache also.